+++ src/main/kotlin/minesweeper/view/render/impl/AdjacentMineCountRenderingStrategy.kt
@@ -1,9 +1,9 @@
-package minesweeper.view.reder.impl
+package minesweeper.view.render.impl
 
 import minesweeper.model.board.Board
 import minesweeper.model.point.Attribute
 import minesweeper.model.point.Coordinate
-import minesweeper.view.reder.MineRenderingStrategy
+import minesweeper.view.render.MineRenderingStrategy
 
 class AdjacentMineCountRenderingStrategy(
     private val board: Board,
